rename & add docs to batchResponseWriter
